+++ example/lib/home_page.dart
@@ -1,4 +1,3 @@
-import 'package:awesome_dialog/awesome_dialog.dart';
 import 'package:example/ume_switch.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
@@ -59,27 +58,6 @@ class _HomePageState extends State<HomePage> {
                 );
               },
               child: const Text('Show Dialog'),
-            ),
-            TextButton(
-              child: const Text('Awesome_dialog show dialog'),
-              onPressed: () {
-                AwesomeDialog(
-                  context: context,
-                  dialogType: DialogType.INFO_REVERSED,
-                  borderSide: BorderSide(color: Colors.green, width: 2),
-                  width: 280,
-                  buttonsBorderRadius: BorderRadius.all(
-                    Radius.circular(2),
-                  ),
-                  headerAnimationLoop: false,
-                  animType: AnimType.BOTTOMSLIDE,
-                  title: 'INFO',
-                  desc: 'Dialog description here...',
-                  showCloseIcon: true,
-                  btnCancelOnPress: () {},
-                  btnOkOnPress: () {},
-                )..show();
-              },
             ),
             TextButton(
               onPressed: () {