+++ kits/flutter_ume_kit_ui/lib/components/widget_detail_inspector/search_bar.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:flutter_ume_kit_ui/util/binding_ambiguate.dart';
 import 'icon.dart' as icon;
 
 typedef void OnSubmitHandle(String text);
@@ -114,7 +115,7 @@ class _SearchInputState extends State<SearchBar> {
     }
     return GestureDetector(
       onTap: () {
-        WidgetsBinding.instance!.addPostFrameCallback((_) {
+        bindingAmbiguate(WidgetsBinding.instance)!.addPostFrameCallback((_) {
           _inputController.clear();
           _focus();
         });