+++ lib/flutter_ume.dart
@@ -7,6 +7,3 @@ export 'core/ui/global.dart';
 export 'service/inspector/inspector_overlay.dart';
 export 'service/vm_service/service_mixin.dart';
 export 'service/vm_service/service_wrapper.dart';
-
-//TODO remove this after no longer support before flutter 3.0.0
-export 'util/binding_ambiguate.dart' show bindingAmbiguate;